Moved keydown keyboard shortcuts app handler to BeakerNotebook, while the actual keyboard navigation occurs on lib/UINotebook
