+++ beaker-vue/src/SingleCellPlayground.vue
@@ -14,9 +14,6 @@
       <Cell 
         :cell="cell"
         class="selected"
-        :selectCell="selectCell"
-        :selectedCellIndex="selectedCellIndex"
-        :index="0"
       />
 
       <br />
